arm_vectors: Extend except vector array to handle all except sources
The exception vetor table was prepared to run software on a single
exception level. In case there is a change of exception level
and interrupt happens there were no handler for that purpose.

The commit adds missing handler entries in the exception vector
table. Note that handlers were not changes and they implementation
may not suite actual needs.

Exception handles will be updated if required in future commit.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>