boot: Remove duplicated code disabling float-point in CPACR_EL1
There was a comment that indicated that execution of instructions
that access Advanced SIMD or Floating-point registers are trapped
to EL1 in case of a Standalone BST.

Actual code what the same if it was build for Free RTOS BPS or
Stanadalone BSP. Also trapping in both cases was disabled.

The commit corrects sources to avoid duplication and updates
the comment to be aligned with code.

Signed-off-by: Piotr Pryga <piotr.pryga@nordicsemi.no>