boot: Add ret to user mode asm function
Add an assembler function that handles switch of exception
level to user mode (EL0). The function should allow to
pass target address in x0 register, a first argument.

There were missing initialization of a timer control
in EL1. It caused and exception when used sleep() function
in a code.

Add missing configuration that disables traping of access
to timer control registers.

In future this may be changed e.g. in case the timer is
used to control precesses/threads scheduling.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>