boot: make reset vector to be separate from vector  table
Extract reset vector from vectors table. The change is a first
step to add other exception vectors handlers to vector table
to enable support of exceptions from other exception levels
also.

This is a step in more general change to make standalone BSP
to use more than single exception level. Xilinx implementation
uses EL3 or EL1 exception only, depending on build configuration.

The extracted reset vector is stored in RVBAR register.
It is also put at address 0x0 to be run at power up.

The consequence of that is move of vector table to be aligned
to an address that has bits[10:0] zeroed. In VBAR_ELx register
these bits are reserved for future use.

Signed-off-by: Piotr Pryga <pryga.piotr@outlook.com>